Add PostgreSQL support.
This commit makes several changes primarily focused on adding PostgreSQL support to the Docker setup. It includes necessary packages in Dockerfile to support PostgreSQL as a database. Additional environment variables for database access are added to the `docker-compose.yml`.

A new requirement, `psycopg2` is added to `requirements.txt` for better PostgreSQL compatibility. The volumes in `docker-compose.yml` have been uncommented,  enabling persistence of data across multiple runs of the containers.