Added command to wait for database and switched to PostgreSQL
This commit adds a new 'wait_for_db' Django management command in 'wait_for_db.py', which stalls the execution of the app until the database is available. This is particularly useful to avoid connection issues in the context of Dockerized services where the database may be temporarily unavailable.

Additionally, the DB engine has been switched from SQLite to PostgreSQL, requiring modifications in DATABASES dictionary in 'settings.py'. Also, a line has been altered in 'checks.yml' indicating the addition of this new command.