Commit message:
Add Recipe model and register it in admin

Implemented a new `Recipe` model in `models.py` that containers necessary fields for a recipe, including a foreign key to the user. Additionally, registered this model in the `admin.py` file to be managed in Django Admin. Also included test cases for creating a new recipe in `test_models.py`. These changes were made to extend our application's functionality to support recipes.