Add routing and views for Recipe API
Included URL routing setup for the Recipe API in the urls.py file. Registered the 'recipes' path in the DefaultRouter for automatic URL generation. In app/urls.py, included the urls from recipe app. Also, created a RecipeSerializer in serializers.py for handling Recipe objects. Modified views.py to add a RecipeViewSet class for managing the recipe APIs and for fetching recipes for an authenticated user. This setup is needed to facilitate CRUD operations for the Recipe model via REST API.